Alexia 26062020 Ole

+++ BusinessEntities/BEDocumentLine.cs
@@ -19,11 +19,11 @@ namespace BusinessEntities
         public string ItemCode { get; set; }
         public string ItemName { get; set; }
         public string LastPurCur { get; set; }
-        public int OnHand { get; set; }
+        public decimal OnHand { get; set; }
         public decimal NumInSale { get; set; }
         public decimal NumInBuy { get; set; }
-        public int Quantity { get; set; }
-        public int _Quantity { get; set; }
+        public decimal  Quantity { get; set; }
+        public decimal _Quantity { get; set; }
         public decimal PriceBefDi { get; set; }
         public decimal DiscPrcnt { get; set; }
         public decimal Price { get; set; }
